+++ src/main/kotlin/bridge/sentence/ErrorSentence.kt
@@ -1,4 +1,4 @@
-package bridge.enum
+package bridge.sentence
 
 enum class ErrorSentence(private val sentence: String) {
     BRIDGE_LENGTH("[ERROR] 다리 길이는 3부터 20 사이의 숫자여야 합니다."),
